Create license
Using the BSD-ISC license, since it's simple (and I've got better things to do than go to court over someone violating the LGPL or Artistic License on my Euchre project).  Sure, it'd be nice to have a license that requires changes to be made available, but it's not like I'm expecting this to be used in commercial contexts and it's not like a license will stop individuals from doing anything.